Story/refactor

+++ src/OzonRoute.Api/Bll/Models/GoodModel.cs
@@ -1,10 +1,10 @@
 namespace OzonRoute.Api.Bll.Models;
 
 public record GoodModel (
-    //In mm
-    int Lenght = 0,
-    int Width = 0,
-    int Height = 0,
+    //In cm
+    double Lenght = 0,
+    double Width = 0,
+    double Height = 0,
     //In kg
     double Weight = 0.0
 ) {}
